dev-requirements.txt
opentelemetry-api==1.0.0
opentelemetry-instrumentation
opentelemetry-exporter-zipkin
opentelemetry-propagator-b3
opentelemetry-exporter-otlp==1.0.0
opentelemetry-instrumentation-aiohttp_client
opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-mysql
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-grpc
opentelemetry-util-http
google
protobuf
pyyaml
pylint==2.4.4
tox
pdoc3
